refactor: Move is_window_function and is_aggregate_function to scratch.py
